+++ Changelog.md
@@ -9,6 +9,7 @@ _yyyy.mm.dd_
 
 ### Bug fix
 * Fix non-determinism issue in ClassificationLearningAgent.
+* Fix intron detection when importing a dot file with TPGGraphDotImporter.
 
 ## Release version 0.2.0
 _2020.06.03_